Drop attaching Gradle Scans ids to reports
The support for this feature would require migration to use new Develocity API. As possibility of browsing scans of single builds happened to be not used at all at Automattic, I think it's not really that useful to justify working on the migration. Instead, this commit simply removes feature of Gradle Scans ids.